+++ Source/Examples/Xamarin.Mac/SimpleDemo/Main.cs
@@ -1,18 +1,13 @@
-﻿using System;
-
-using AppKit;
+﻿using AppKit;
 
 namespace SimpleDemo
 {
     static class MainClass
     {
-        static void Main (string[] args)
+        static void Main(string[] args)
         {
-            NSApplication.Init ();
-
-            var application = NSApplication.SharedApplication;
-            application.Delegate = new AppDelegate ();
-            application.Run ();
+            NSApplication.Init();
+            NSApplication.Main(args);
         }
     }
 }