+++ Source/OxyPlot.Xamarin.Android/Properties/AssemblyInfo.cs
@@ -1,0 +1,25 @@
+﻿using System.Reflection;
+using System.Runtime.CompilerServices;
+
+// Information about this assembly is defined by the following attributes. 
+// Change them to the values specific to your project.
+
+[assembly: AssemblyTitle("OxyPlot.Xamarin.Android")]
+[assembly: AssemblyDescription("OxyPlot for Xamarin.Android")]
+[assembly: AssemblyConfiguration("")]
+[assembly: AssemblyCompany("OxyPlot")]
+[assembly: AssemblyProduct("OxyPlot")]
+[assembly: AssemblyCopyright("Copyright (c) 2014 OxyPlot contributors")]
+[assembly: AssemblyTrademark("")]
+[assembly: AssemblyCulture("")]
+
+// The version numbers are updated by the build script. See ~/appveyor.yml
+[assembly: AssemblyVersion("0.0.1")]
+[assembly: AssemblyInformationalVersion("0.0.1-alpha")]
+[assembly: AssemblyFileVersion("0.0.1")]
+
+// The following attributes are used to specify the signing key for the assembly, 
+// if desired. See the Mono documentation for more information about signing.
+
+//[assembly: AssemblyDelaySign(false)]
+//[assembly: AssemblyKeyFile("")]