+++ Source/OxyPlot.Xamarin.Forms.Platform.UWP/Properties/AssemblyInfo.cs
@@ -6,24 +6,17 @@ using System.Runtime.InteropServices;
 // set of attributes. Change these attribute values to modify the information
 // associated with an assembly.
 [assembly: AssemblyTitle("OxyPlot.Xamarin.Forms.Platform.UWP")]
-[assembly: AssemblyDescription("")]
+[assembly: AssemblyDescription("OxyPlot for Windows Universal (UWP)")]
 [assembly: AssemblyConfiguration("")]
-[assembly: AssemblyCompany("")]
-[assembly: AssemblyProduct("OxyPlot.Xamarin.Forms.Platform.UWP")]
-[assembly: AssemblyCopyright("Copyright ©  2015")]
+[assembly: AssemblyCompany("OxyPlot")]
+[assembly: AssemblyProduct("OxyPlot")]
+[assembly: AssemblyCopyright("Copyright (c) 2014 OxyPlot contributors")]
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]
 
-// Version information for an assembly consists of the following four values:
-//
-//      Major Version
-//      Minor Version 
-//      Build Number
-//      Revision
-//
-// You can specify all the values or you can default the Build and Revision Numbers 
-// by using the '*' as shown below:
-// [assembly: AssemblyVersion("1.0.*")]
-[assembly: AssemblyVersion("1.0.0.0")]
-[assembly: AssemblyFileVersion("1.0.0.0")]
+// The version numbers are updated by the build script. See ~/appveyor.yml
+[assembly: AssemblyVersion("0.0.1")]
+[assembly: AssemblyInformationalVersion("0.0.1-alpha")]
+[assembly: AssemblyFileVersion("0.0.1")]
+
 [assembly: ComVisible(false)]