+++ CHANGELOG.md
@@ -8,6 +8,8 @@ All notable changes to this project will be documented in this file.
 
 ### Removed
 - OxyPlot.MonoTouch (#50)
+- OxyPlot.Xamarin.Forms.Platform.WP8 (#74)
+- OxyPlot.MonoMac (#49)
 
 ## 1.0.0 - 2017-01-08
 ### Added